add loging from init
