+++ .env.example
@@ -2,6 +2,8 @@ APP_VERSION=0.0.1
 
 APP_LOG_LEVEL=DEBUG
 LOG_FILE_BACKUP_COUNT=10
+LOGFILE_PATH =/home/etdadm/logs/etd/etd_dash.log
 
 CONSUME_QUEUE_NAME: etd_submission_ready
 PUBLISH_QUEUE_NAME: etd_in_storage
+BROKER_URL: ""
